:arrow_up_small: move thenable<->promise utility to async
